Combine code duplication of `StaticArrayQueue` and `ArrayQueue`
